ESIHO-0.20

+++ core/src/com/esiho/world/item/Arme.java
@@ -28,7 +28,7 @@ public class Arme extends Item {
         return team;
     }
 
-    public int getCoeffDegats(){
+    public double getCoeffDegats(){
         return type.coeffDegats;
     }
 }
